backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
langchain==0.2.5
langchain-community==0.2.5
langchain-google-genai==1.0.6
faiss-cpu==1.8.0
rank-bm25==0.2.2
pypdf==4.2.0
google-generativeai==0.5.4
python-dotenv==1.0.1
langchain-huggingface==0.0.3
sentence-transformers==3.0.1